Fix bug on relocation of R_X86_64_PC32 in rank-prelude
This is a linker error that occurs when you try to create a shared object from a static library that was not compiled with the -fPIC flag. This flag tells the compiler to generate position-independent code, which is required for shared objects. To fix this error, the rank-prelude library was made to be statically compiled, now